feat: add JWT middleware and protect property routes

+++ src/entities/property/api/deleteProperty.ts
@@ -1,6 +1,15 @@
 export const deleteProperty = async (id: string) => {
+	const token = localStorage.getItem('token');
+
+	if (!token) {
+		throw new Error('No token found. Please login.');
+	}
+
 	const res = await fetch(`http://localhost:5000/api/properties/${id}`, {
 		method: 'DELETE',
+		headers: {
+			Authorization: `Bearer ${token}`, // 🔐 ОБОВʼЯЗКОВО
+		},
 	});
 
 	if (!res.ok) {
